### 2024-06-13 #155 @editorialbot
[PRE REVIEW]: RobustNeuralNetworks.jl: a Package for Machine Learning and Data-Driven Control with Certified Robustness
**Submitting author:** @nic-barbara (<a href="http://orcid.org/0000-0002-4885-3307">Nicholas Barbara</a>)
**Repository:** https://github.com/acfr/RobustNeuralNetworks.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.2
**Managing EiC:** Mathieu Besançon

Thanks for submitting your paper to JCON @nic-barbara. **Currently, there isn't a JCON editor assigned** to your paper.
