### 2024-07-09 #155 @lucaferranti
Hi @nic-barbara ,

I found one reviewer and looking for a second one.

Meanwhile, I noticed some of the references don't have DOIs, could you check if some of those have a DOI available that could be added/
