### 2024-07-09 #163 @editorialbot
[REVIEW]: RobustNeuralNetworks.jl: a Package for Machine Learning and Data-Driven Control with Certified Robustness
**Submitting author:** @nic-barbara (<a href="http://orcid.org/0000-0002-4885-3307">Nicholas Barbara</a>)
**Repository:** https://github.com/acfr/RobustNeuralNetworks.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.4
