### 2024-07-09 #163 @lucaferranti
Hi @asinghvi17 and @pevnak :wave: ,

thank you very much for volunteering as reviewers. I will be the editor for this submission, feel free to ping me if you have any questions. Also maake sure to check the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/)

You can generate your reviewer checklist with

As you go through the checklist and review the paper, you can either leave comments here or open issues in the linked repository.
