### 2024-08-09 #163 @pevnak
I do not know, where I can find the `paper.tex` file. I have looked at https://arxiv.org/pdf/2306.12612 which has the affiliations and references. Most of them has DOI, but those referencing to Arxiv naturally don't.

For me, this is very good contribution. I have tried the package and it works as advertised. I have skim-read the papers describing the implemented methods and love them.

I have not find the statement of need, but the motivation in papers, documentations, and implemented work is more than enough. Moreover, any person who is interested enough in fundamental principles of NNs (and machine learning) would value this work. I am excited about this.
