### 2024-08-09 #163 @lucaferranti
> I do not know, where I can find the paper.tex file.

@pevnak thank you for highlighting this! So, in JuliaCon proceedings the paper lives in a `paper` folder and we allow two options: 1. to have that on the main branch 2. to have that on a separate branch. The second option (which is the case here), is particularly popular if the submission is a package and the authors want to keep the main branch minimal. Long story short, if you go to the repository, there is a `paper` branch which has the `paper/paper.tex` .

**However** actually authors names and affiliations should **not** be in the `paper.tex`. They are in a separate `header.tex` which is automatically generaed from the `paper.yml` (this is because the bot needs a `paper.yml`). That checkbox is hence inaccurate. Thank you very much for bringing this to my attention, I'll fix it in the checklist template!. For that checkbox, it is enough to check that the final paper pdf has a correct list of authors and affiliations.
