### 2024-08-09 #163 @lucaferranti
> I have not find the statement of need, but the motivation in papers, documentations, and implemented work is more than enough.

Yet another thing I need to fix on the checklist template :sweat_smile: . The important thing is that the need for the presented work is clearly conveyed from the paper. A section with the title "Statement of Need" verbatim is not required
