### 2024-08-10 7d27602
Minor edits to documentation
