### 2024-08-10 #163 @nic-barbara
Thanks for the review @pevnak, much appreciated! I've taken care of the small bug in the documentation you pointed out too.

@lucaferranti regarding DOIs, as far as I'm aware the citations with missing DOIs don't actually have a DOI that I can add in. I've included one wherever possible.
