### 2024-08-10 2cbdedf
Minor updates to references
