+++ paper/ref.bib
@@ -13,17 +13,20 @@
    volume = {67},
    year = {2022},
 }
-@article{Pauli++2022c,
-   abstract = {In this work, we propose a dissipativity-based method for Lipschitz constant estimation of 1D convolutional neural networks (CNNs). In particular, we analyze the dissipativity properties of convolutional, pooling, and fully connected layers making use of incremental quadratic constraints for nonlinear activation functions and pooling operations. The Lipschitz constant of the concatenation of these mappings is then estimated by solving a semidefinite program which we derive from dissipativity theory. To make our method as efficient as possible, we exploit the structure of convolutional layers by realizing these finite impulse response filters as causal dynamical systems in state space and carrying out the dissipativity analysis for the state space realizations. The examples we provide show that our Lipschitz bounds are advantageous in terms of accuracy and scalability.},
-   author = {Patricia Pauli and Dennis Gramlich and Frank Allgöwer},
-   journal = {Proceedings of Machine Learning Research},
-   keywords = {Convolutional neural networks,dissipativity,incremental quadratic con-straints,robustness},
-   month = {11},
-   pages = {1-14},
-   title = {Lipschitz constant estimation for 1D convolutional neural networks},
-   volume = {211},
-   url = {https://arxiv.org/abs/2211.15253v2},
-   year = {2022},
+@InProceedings{Pauli++2022c,
+  title = 	 {Lipschitz constant estimation for 1D convolutional neural networks},
+  author =       {Pauli, Patricia and Gramlich, Dennis and Allg\"ower, Frank},
+  booktitle = 	 {Proceedings of The 5th Annual Learning for Dynamics and Control Conference},
+  pages = 	 {1321--1332},
+  year = 	 {2023},
+  editor = 	 {Matni, Nikolai and Morari, Manfred and Pappas, George J.},
+  volume = 	 {211},
+  series = 	 {Proceedings of Machine Learning Research},
+  month = 	 {15--16 Jun},
+  publisher =    {PMLR},
+  pdf = 	 {https://proceedings.mlr.press/v211/pauli23a/pauli23a.pdf},
+  url = 	 {https://proceedings.mlr.press/v211/pauli23a.html},
+  abstract = 	 {In this work, we propose a dissipativity-based method for Lipschitz constant estimation of 1D convolutional neural networks (CNNs). In particular, we analyze the dissipativity properties of convolutional, pooling, and fully connected layers making use of incremental quadratic constraints for nonlinear activation functions and pooling operations. The Lipschitz constant of the concatenation of these mappings is then estimated by solving a semidefinite program which we derive from dissipativity theory. To make our method as efficient as possible, we exploit the structure of convolutional layers by realizing these finite impulse response filters as causal dynamical systems in state space and carrying out the dissipativity analysis for the state space realizations. The examples we provide show that our Lipschitz bounds are advantageous in terms of accuracy and scalability.}
 }
 @article{Junnarkar++2023,
    abstract = {We propose a parameterization of a nonlinear dynamic controller based on the recurrent equilibrium network, a generalization of the recurrent neural network. We derive constraints on the parameterization under which the controller guarantees exponential stability of a partially observed dynamical system with sector bounded nonlinearities. Finally, we present a method to synthesize this controller using projected policy gradient methods to maximize a reward function with arbitrary structure. The projection step involves the solution of convex optimization problems. We demonstrate the proposed method with simulated examples of controlling a nonlinear inverted pendulum.},
