+++ v0.3.3/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"introduction/developing/#Contributing-to-the-Package","page":"Contributing to the Package","title":"Contributing to the Package","text":"","category":"section"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"All contributors welcome! Please contact nicholas.barbara@sydney.edu.au with any questions.","category":"page"},{"location":"introduction/developing/#Installation-for-Development","page":"Contributing to the Package","title":"Installation for Development","text":"","category":"section"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"If you would like to contribute the package, please clone the repository into your ~/.julia/dev/ directory with","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"git clone git@github.com:acfr/RobustNeuralNetworks.jl.git RobustNeuralNetworks","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"info: The name matters\nNote that the repository is named RobustNeuralNetworks.jl but your folder should be named RobustNeuralNetworks (no .jl). This is convention for Julia packages, and will ensure the package manager knows where to look.Also make sure you have cloned the repo inside your ~/.julia/dev/ folder for the following instructions to work. If the dev/ directory does not exist, create it. ","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"Navigate to ~/.julia/dev/RobustNeuralNetworks/, start a Julia session, and type the following in the REPL to activate the package.","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"using Pkg\nPkg.activate(\".\")\nPkg.instantiate()","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"Check that the example in Getting Started runs without errors and matches the given output before continuing.","category":"page"},{"location":"introduction/developing/#Git-Workflow","page":"Contributing to the Package","title":"Git Workflow","text":"","category":"section"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"The package is just another git repository, so the development workflow is similar to most projects.","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"When developing features, please create a new branch labelled feature/<some_descriptive_name>. For example, the branch feature/documentation is where this documentation was first written and tested.\nIf you notice a bug, please create a git issue and a new branch associated with that issue to let others know what you are working on. The branch should be deleted once the issue is closed.\nSubmit pull requests once you have completed a new feature and tested it thorougly. All pull requests must be approved by at least one other developer of the package, and must pass the continuous integration pipeline.","category":"page"},{"location":"introduction/developing/#Package-Structure","page":"Contributing to the Package","title":"Package Structure","text":"","category":"section"},{"location":"introduction/developing/#Layout","page":"Contributing to the Package","title":"Layout","text":"","category":"section"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"The package is divided into four main directories:","category":"page"},{"location":"introduction/developing/","page":"Contributing to the Package","title":"Contributing to the Package","text":"The src/ folder contains all source code required f
[…]
