### 2024-08-10 #163 @lucaferranti
@pevnak @nic-barbara For the DOIs, the rule of thumb is "DOIs should be included whenever possible" You can use `editorialbot` to check the references DOI status (as I just did). The summary can be interpreted as follows:

- **Invalid DOIs**: They indicate typos/misformatting and should be fixed
- **Missing DOIs**: It's good to check that if a DOI can be found, note that sometimes the bot will also suggest a DOI for it if it can find one from the internet (not in this case). If those have no DOI, then it's of course ok. I.e. the missing DOIs section in more of a suggestion to double check, but not a strict requirement
