### 2024-08-12 #163 @pevnak
I think the paper is missing a section Statement of need. Not sure it is needed, as from introduction and examples, this is needed. Also, I have not seen similar methods implemented elsewhere.
