### 2024-08-27 #163 @nic-barbara
Hi @lucaferranti, @pevnak, @asinghvi17, just wondering if there's anything I can do to expedite this review process? As far as I'm aware there's nothing I need to change so far: it seems that an explicit statement of need section is not required according to @lucaferranti, and the papers with missing DOIs don't actually have DOIs. Is this correct?
