### 2024-08-28 #163 @lucaferranti
> it seems that an explicit statement of need section is not required according to @lucaferranti, and the papers with missing DOIs don't actually have DOIs. Is this correct?

It is important for the paper to clearly convey its motivation, but this can be done e..g as part of the introduction or other parts of the paper. A section called "statement of need" is not required, this has been recently updated on hte checklist

> the papers with missing DOIs don't actually have DOIs

correct, the check is more of a fyi guideline and not assertive. If you have checked that those papers don't have a DOI, then it's ok. In a lot of cases the bot is smart enough to find and suggest a potential DOI, since it doens't find any in this case, it is likely that they don't have one
