### 2024-10-16 #163 @lucaferranti
Hi @pevnak :wave: ,

thank you for your careful review. I see you have a couple of boxes unticked. Could you let the authors know if there are some further changes you'd expect to see regarding those? Otherwise, can you confirm the submission is ready from your side?
