### 2025-02-10 06ef70a
minor grammar suggestions
