+++ paper/TexFiles/ExampleMNIST.tex
@@ -53,7 +53,7 @@ model = Chain(DiffLBDN(model_ps), Flux.softmax)
 
 The \verb|model| consists of two parts. The first is a callable \verb|DiffLBDN| model constructed from its direct parameterization, which is defined by an instance of \verb|DenseLBDNParams| as per Section \ref{sec:parameterizations}. The output is then converted to a probability distribution using a \verb|softmax| layer. Note that all \verb|AbstractLBDN| models can be combined with traditional neural network layers using \verb|Flux.Chain|. 
 
-We could also construct the \verb|model| as a chain of \verb|SandwichFC| layers. Introduced in \cite{Wang+Manchester2023}, the ``sandiwch'' layer is a dense layer with a guaranteed Lipschitz bound of 1.0. We have designed the user interface for \verb|SandwichFC| similarly to that of \verb|Flux.Dense|.
+We could also construct the \verb|model| as a chain of \verb|SandwichFC| layers. Introduced in \cite{Wang+Manchester2023}, the ``sandwich'' layer is a dense layer with a guaranteed Lipschitz bound of 1.0. We have designed the user interface for \verb|SandwichFC| similarly to that of \verb|Flux.Dense|.
 \begin{lstlisting}[language = Julia]
 model = Chain(
     (x) -> (sqrt(γ) * x),
