+++ paper/paper.tex
@@ -24,7 +24,7 @@
 \begin{abstract}
 
 Neural networks are typically sensitive to small input perturbations, leading to unexpected or brittle behaviour. 
-We present \verb|RobustNeuralNetworks.jl|: a Julia package for neural network models that are constructed to naturally satisfy a set of user-defined robustness constraints. The package is based on the recently proposed Recurrent Equilibrium Network (REN) and Lipschitz-Bounded Deep Network (LBDN) model classes, and is designed to interface directly with Julia's most widely-used machine learning package, \verb|Flux.jl|. We discuss the theory behind our model parameterization, give an overview of the package, and provide a tutorial demonstrating its use in image classification, reinforcement learning, and nonlinear state-observer design.
+We present \verb|RobustNeuralNetworks.jl|: a Julia package for neural network models that are constructed to naturally satisfy a set of user-defined robustness metrics. The package is based on the recently proposed Recurrent Equilibrium Network (REN) and Lipschitz-Bounded Deep Network (LBDN) model classes, and is designed to interface directly with Julia's most widely-used machine learning package, \verb|Flux.jl|. We discuss the theory behind our model parameterization, give an overview of the package, and provide a tutorial demonstrating its use in image classification, reinforcement learning, and nonlinear state-observer design.
 
 \end{abstract}
 
