+++ paper/TexFiles/Conclusions.tex
@@ -1,4 +1,4 @@
-This paper has presented \verb|RobustNeuralNetworks.jl|, a Julia package for robust machine learning based on the recently-proposed Recurrent Equilibrium Network (REN) and Lipschitz-Bounded Deep Network (LBDN) model classes. The models are unique in that they naturally satisfy a set of \textit{built-in} robustness certificates, such as contraction and Lipschitz bounds. We have presented an overview of the model architectures, including background theory on robustness metrics in nonlinear systems, and have outlined the package structure and its usage alongside Julia's main machine-learning library, \verb|Flux.jl|. We have demonstrated via examples in image classification, reinforcement learning, and observer design that the package is easy to use in many common machine learning and data-driven control problems, while also offering the advantage of robustness guarantees.
+This paper has presented \verb|RobustNeuralNetworks.jl|, a Julia package for robust machine learning based on the recently-proposed Recurrent Equilibrium Network (REN) and Lipschitz-Bounded Deep Network (LBDN) model classes. The models are unique in that they naturally satisfy a set of \textit{built-in} robustness metrics, such as contraction and Lipschitz bounds. We have presented an overview of the model architectures, including background theory on robustness metrics in nonlinear systems, and have outlined the package structure and its usage alongside Julia's main machine-learning library, \verb|Flux.jl|. We have demonstrated via examples in image classification, reinforcement learning, and observer design that the package is easy to use in many common machine learning and data-driven control problems, while also offering the advantage of robustness guarantees.
 
 We intend \verb|RobustNeuralNetworks.jl| to be widely-applicable in the scientific and machine learning communities for learning-based problems in which robustness certificates are crucial, and have already used the package in our own research in robust reinforcement learning \cite{Barbara++2023}. Some areas in which this package will be most applicable include: data-driven control and state estimation, image classification and segmentation, and privacy and security. We intend to expand the package with more robust neural network architectures in the future. Examples include LBDNs with one-dimensional convolution \cite{Pauli++2022c} and circular convolutions \cite{Wang+Manchester2023}, continuous-time REN models \cite{Martinelli++2023}, and RENs respecting other non-Euclidean contraction metrics \cite{Davydov++2022}. We encourage any and all contributions to \verb|RobustNeuralNetworks.jl| to further its use in robust machine learning problems.
 
