### 2025-02-12 #163 @nic-barbara
Hi @burtonjosh, thank you for your comments! I have taken your recommendations onboard and made edits to the paper via [this PR](https://github.com/acfr/RobustNeuralNetworks.jl/pull/152). I have also merged your PR correcting minor grammatical edits. In response to your more specific comments, please see below.

- Thank you for the suggestion to clarify the wording around robustness certificates. I have re-worded mentions of robustness certificates/metrics throughout the paper to now be consistent. In summary, a robustness metric describes constraints on the internal stability and input-output sensitivity of a network. If a network is *guaranteed* to satisfy a particular robustness metric, we call this a robustness certificate. I have removed mention of robustness constraints/criteria/requirements for clarity.

- The notation $0 \preceq Q$ means that the matrix $Q$ is [positive semi-definite](https://en.wikipedia.org/wiki/Definite_matrix). As it turns out, this was a typo! $Q$ should always be negative semi-definite in this context ($0 \succeq Q$). I have fixed this typo and added a note for clarity.

- I have added a couple of sentences on passivity, and added a couple of citations to show where it can be useful in practice. We have only recently started using passive RENs in our research, so I have not included an example in this paper. We might add something in the package documentation later down the track.

- Great question regarding acyclic LBDNs. If you were to write the LBDN architecture as a REN, it would automatically have an acyclic structure. This is detailed in ["Direct parameterisation of Lipschitz-bounded deep networks" by Wang & Manchester, 2023](https://arxiv.org/abs/2301.11526). I have added a note on this at the end of Remark 1.

- Thank you for your comment on comparing speed against models with no guarantees. This is a great point and something that we'll certainly consider adding to the docs in future iterations of the package. From my own personal experience in Deep RL tasks, training RENs and LBDNs with robustness certificates typically takes about 2-3 times as long as a standard model. However, this (obviously) varies with the size of the network.

While making these edits I have also added a few extra references and tidied up the wording of a few sections. Please let me know if you have any further comments or concerns!
