### 2025-02-12 #163 @lucaferranti
@nic-barbara great, please check the post-review checklist above and the references check in the message above.

- Some references don't have a DOI, if you can find one, that's great please add it. If there is no DOI, that's ok
- To finalize the submission, the software shoudl have a new release with all new changes and the corresponding repository should be archived in zenodo, once this is done, please post version and doi here
