+++ paper/ref.bib
@@ -2,11 +2,11 @@
    abstract = {This paper presents a study of robust policy networks in deep reinforcement learning. We investigate the benefits of policy parameterizations that naturally satisfy constraints on their Lipschitz bound, analyzing their empirical performance and robustness on two representative problems: pendulum swing-up and Atari Pong. We illustrate that policy networks with small Lipschitz bounds are significantly more robust to disturbances, random noise, and targeted adversarial attacks than unconstrained policies composed of vanilla multi-layer perceptrons or convolutional neural networks. Moreover, we find that choosing a policy parameterization with a non-conservative Lipschitz bound and an expressive, nonlinear layer architecture gives the user much finer control over the performance-robustness trade-off than existing state-of-the-art methods based on spectral normalization.},
    author = {Nicholas H. Barbara and Ruigang Wang and Ian R. Manchester},
    isbn = {2405.11432v1},
-   journal = {arXiv preprint arXiv:2405.11432},
    keywords = {Lipschitz networks,reinforcement learning,robustness},
    month = {5},
    title = {On Robust Reinforcement Learning with Lipschitz-Bounded Policy Networks},
    url = {https://arxiv.org/abs/2405.11432v1},
+   doi = {10.48550/arXiv:2405.11432},
    year = {2024},
 }
 @article{Kucera1975,
@@ -27,10 +27,11 @@
       title={Learning Stable and Passive Neural Differential Equations}, 
       author={Jing Cheng and Ruigang Wang and Ian R. Manchester},
       year={2024},
-      eprint={2404.12554},
+      journal = {arXiv preprint},
       archivePrefix={arXiv},
       primaryClass={eess.SY},
       url={https://arxiv.org/abs/2404.12554}, 
+      doi={10.48550/arXiv.2404.12554}
 }
 @book{vanderSchaft2017,
    author = {Arjan van der Schaft},
