### 2025-02-13 #163 @nic-barbara
Thanks @lucaferranti! Confirming that I have:

- Checked for DOIs on the references. I found two more and added them.
- Double-checked authors and affiliations, including ORCIDs.
- Made a new release of the software with the latest changes from the review. This is [`v0.3.4`](https://github.com/acfr/RobustNeuralNetworks.jl/releases/tag/v0.3.4).
- Made a Zenodo release. DOI is: [10.5281/zenodo.14861888](https://doi.org/10.5281/zenodo.14861888).
- Checked that title and author list (including ORCIDs) match the paper.
- Checked that both the code repo and Zenodo archive has the same license (MIT license).
