+++ paper/ref.bib
@@ -6,7 +6,7 @@
    month = {5},
    title = {On Robust Reinforcement Learning with Lipschitz-Bounded Policy Networks},
    url = {https://arxiv.org/abs/2405.11432v1},
-   doi = {10.48550/arXiv:2405.11432},
+   doi = {10.48550/arXiv.2405.11432},
    year = {2024},
 }
 @article{Kucera1975,
