### 2025-02-13 #163 @nic-barbara
Thanks @lucaferranti! I just fixed that one broken DOI, it's now `10.48550/arXiv.2405.11432` not `10.48550/arXiv:2405.11432`. With this fix, I'm happy for you to finalise it!
